+++ src/settings/NumberGeneratorSequenceRoute/SequenceView/SequenceView.test.js
@@ -1,13 +1,13 @@
-import { render, screen, waitFor } from '@folio/jest-config-stripes/testing-library/react';
+import { waitFor } from '@folio/jest-config-stripes/testing-library/react';
 import { Field as MockField } from 'react-final-form';
 
 import { TextField as MockTextField } from '@folio/stripes/components';
 
-import { Button, Callout, KeyValue, TextField, renderWithIntl } from '@folio/stripes-erm-testing';
+import { Button, Callout, KeyValue, TextField } from '@folio/stripes-erm-testing';
 
 import SequenceView from './SequenceView';
-import { translationsProperties } from '../../../test/helpers';
-import { numberGenerator1, numberGenerator2 } from '../../../test/jest/mockGenerators';
+import { numberGenerator1, numberGenerator2 } from '../../../../test/jest/mockGenerators';
+import { renderWithTranslations } from '../../../../test/helpers';
 
 const onClose = jest.fn();
 const mockSequence = numberGenerator1?.sequences[0];
@@ -15,8 +15,8 @@ const mockGenerators = [numberGenerator1, numberGenerator2];
 
 const fakeCalloutInfo = { id: '123', label: 'numgenName', code: 'numgenCode' };
 
-jest.mock('../../public', () => ({
-  ...jest.requireActual('../../public'),
+jest.mock('../../../public', () => ({
+  ...jest.requireActual('../../../public'),
   useNumberGeneratorSequence: jest.fn(() => ({ data: mockSequence })),
   useMutateNumberGeneratorSequence: ({ afterQueryCalls: { delete: deleteQueryCalls, put: putQueryCalls } }) => ({
     put: () => Promise.resolve(true).then(() => putQueryCalls(mockGenerators[0], fakeCalloutInfo)),
@@ -37,7 +37,7 @@ jest.mock('../../public', () => ({
   });
 }); */
 
-jest.mock('./NumberGeneratorSequenceForm', () => () => {
+jest.mock('../NumberGeneratorSequenceForm', () => () => {
   return (
     <>
       NumberGeneratorSequenceForm
@@ -54,13 +54,11 @@ jest.mock('./NumberGeneratorSequenceForm', () => () => {
 let renderComponent;
 describe('SequenceView', () => {
   beforeEach(async () => {
-    renderComponent = renderWithIntl(
+    renderComponent = renderWithTranslations(
       <SequenceView
         match={{ params: { seqId: mockSequence?.id } }}
         onClose={onClose}
       />,
-      translationsProperties,
-      render,
       // Ensure right intl key is used in kint-comps
       {
         intlKey: 'ui-service-interaction',
@@ -82,7 +80,7 @@ describe('SequenceView', () => {
   });
 
   test('renders expected sequence checkDigitAlgo', async () => {
-    await KeyValue('Checksum').has({ value: 'EAN13' });
+    await KeyValue('Method').has({ value: '31-RTL-mod10-I (EAN)' });
   });
 
   test('renders expected sequence enabled', async () => {