+++ test/helpers/index.js
@@ -1,1 +1,2 @@
 export { default as translationsProperties } from './translationsProperties';
+export { default as renderWithTranslations } from './renderWithTranslations';